Ameliorations de la partie avis avec approbation

+++ src/main/java/club/taekwondo/repository/jpa/AvisRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvisRepository extends JpaRepository<Avis, Long> {
-    // Ici, tu peux ajouter des méthodes spécifiques si nécessaire
+public interface AvisRepository extends JpaRepository<Avis, Integer> {
+    // Ici, vous pouvez ajouter des méthodes spécifiques si nécessaire
 }
